idle-detection/interceptor.https.html

<!DOCTYPE html>
<link rel="help" href="https://github.com/inexorabletash/idle-detection">
<title>Tests the Idle Detection API</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/mojo/public/mojom/base/string16.mojom.js"></script>
<script src="/gen/mojo/public/mojom/base/time.mojom.js"></script>
<script src="/gen/third_party/blink/public/platform/modules/idle/idle_manager.mojom.js"></script>
<script src="mock.js"></script>
<script>
'use strict';

promise_test(async t => {
  // Basic test that expects start() to call internally
  // addMonitor, which in turn return an ACTIVE state.
  expect(addMonitor).andReturn((threshold, monitorPtr) => {
      return Promise.resolve({
        state: {
          user: UserIdleState.ACTIVE,
          screen: ScreenIdleState.LOCKED
        }
      });
  });

  let status = new IdleDetector({threshold: 10});

  let watcher = new EventWatcher(t, status, ["change"]);

  await status.start();

  // Waits for the first event.
  await watcher.wait_for("change");

  assert_equals(status.state.user, "active");
  assert_equals(status.state.screen, "locked");

  status.stop();
}, 'query()');

promise_test(async t => {
  // Verifies that an event is thrown when a change of state from IDLE to ACTIVE
  // is detected.
  expect(addMonitor).andReturn((threshold, monitorPtr) => {
      let first = Promise.resolve({
        state: {
          user: UserIdleState.ACTIVE,
          screen: ScreenIdleState.UNLOCKED
        }
      });

      t.step_timeout(() => {
        monitorPtr.update({
          user: UserIdleState.IDLE,
          screen: ScreenIdleState.UNLOCKED
        });
      }, 0);

      return first;
    });

  let monitor = new IdleDetector({threshold: 10});

  let watcher = new EventWatcher(t, monitor, ["change"]);

  await monitor.start();

  // Wait for the initial state.
  await watcher.wait_for("change");

  // Wait for the first change in state.
  await watcher.wait_for("change");

  assert_equals(monitor.state.user, "idle");
  assert_equals(monitor.state.screen, "unlocked");

  monitor.stop();
}, 'updates once');


promise_test(async t => {
  // Simulates the user being active, going idle and then going back active
  // again.
  expect(addMonitor).andReturn((threshold, monitorPtr) => {
      let first = Promise.resolve({
        state: {
          user: UserIdleState.ACTIVE,
          screen: ScreenIdleState.UNLOCKED
        }
      });

      // Updates the client once with the user idle.
      t.step_timeout(() => {
        monitorPtr.update({
          user: UserIdleState.IDLE,
          screen: ScreenIdleState.UNLOCKED
        });
      }, 0);
      // Updates the client a second time with the user active.
      t.step_timeout(() => {
        monitorPtr.update({
          user: UserIdleState.ACTIVE,
          screen: ScreenIdleState.UNLOCKED
        });
      }, 1);
      return first;
    });

  let monitor = new IdleDetector({threshold: 10});

  let watcher = new EventWatcher(t, monitor, ["change"]);

  await monitor.start();

  // Waits for the initial state.
  await watcher.wait_for("change");

  // Waits for the first event.
  await watcher.wait_for("change");
  assert_equals(monitor.state.user, "idle");

  // Waits for the second event.
  await watcher.wait_for("change");
  assert_equals(monitor.state.user, "active");

  monitor.stop();
}, 'updates twice');

promise_test(async t => {
  // Simulates a locked screen.
  expect(addMonitor).andReturn((threshold, monitorPtr) => {
      return Promise.resolve({
        state: {
          user: UserIdleState.ACTIVE,
          screen: ScreenIdleState.LOCKED
        }
      });
    });

  let monitor = new IdleDetector({threshold: 10});

  let watcher = new EventWatcher(t, monitor, ["change"]);

  await monitor.start();

  // waits for the initial state.
  await watcher.wait_for("change");

  assert_equals(monitor.state.screen, "locked");

  monitor.stop();
}, 'locked screen');

promise_test(async t => {
  expect(addMonitor).andReturn((threshold, monitorPtr) => {
      return Promise.resolve({
        state: {
          user: UserIdleState.ACTIVE,
          screen: ScreenIdleState.LOCKED
        }
      });
  });

  let status = new IdleDetector({threshold: 10});

  let event = new Promise((resolve, reject) => {
    status.onchange = resolve;
  });

  await status.start();

  // Waits for the first event.
  await event;

  assert_equals(status.state.user, "active");
  assert_equals(status.state.screen, "locked");

  status.stop();
}, 'IdleDetector.onchange');
</script>
